requirements.txt
grpcio
asyncio
numpy